Finished implementing KICK command.
Notes:
Implementing KICK command needed a few changes in the existing code to
keep it clean, The new changes shouldn't effect how the rest of the code
works.

A new numeric reply ( ERR_MULITPLEPARAMS ) that doesn't exist in RFC2119 protocol was added,
it uses ERR_UNKNOWNERROR (400) to let the client know that the server
doesn't support having mulitiple targets.

This version of code isn't tested yet since the command was implemented
before the JOIN command was added to the code.

Changes:
    - Reimplemented KICK command to match the new version of the code.
    - KICK command doesn't support mulitiple targets anymore.
    - Added the new system to send numeric replies back to the client,
      where all the replies are saved in the replies.hpp file
    - Added a utils namespace that has common utils functions and
      used them in the existing code.
    - Added getOrigin() method to AClient class to make constructing
      messages easier.